% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{campbell}
\alias{campbell}
\title{Raw Vortex outputs from Campbell et al (2016)}
\format{Several .dat and .stdat files.}
\source{
Campbell et al. (2016). Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management, 23, 81-99
}
\usage{
# To retrieve the path to the files use:
# system.file("extdata", "campbell", package="vortexR")
}
\description{
A folder with Vortex outputs from Campbell et al (2016).
}
