% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vortexr.R
\name{CompileIter}
\alias{CompileIter}
\title{Compile iterations from one .yr file}
\usage{
CompileIter(iter, filename, n_rows, iter_ln, lines, header)
}
\arguments{
\item{iter}{The iteration (run) number}

\item{filename}{The fully qualified filename to read from}

\item{n_rows}{The number of rows to read from the file}

\item{iter_ln}{The number of rows to skip from the file}

\item{lines}{An object returned from readLines()}

\item{header}{A character vector of column names}
}
\value{
A data.frame
}
\description{
Compile iterations from one .yr file and add a column with scenario names and
one with iteration number
}

