% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexr.R
\name{pac.clas.Nadults}
\alias{pac.clas.Nadults}
\title{Harmonic mean of adults and population sizes}
\format{A \code{data.frame} with 24 observations of 4 variables.}
\source{
Pacioni, C., Spencer, P.B.S., Lacy, R.C., and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas.Nadults)
}
\description{
Data from Pacioni et al. (in prep) - sensitivity test scenario
"ST_Classic" - were used to calculate the harmonic mean of adults and population
sizes using \code{Nadults}.
}

