% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.clas.Ne}
\alias{pac.clas.Ne}
\title{Effective population size}
\format{A \code{data.frame} with 24 observations of 2 variables.}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas.Ne)
}
\description{
Data from Pacioni et al. (2017) - sensitivity test scenario
"ST_Classic" - were used to calculate the effective population size
sizes using \code{Ne}.
}
