% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.clas.pairw}
\alias{pac.clas.pairw}
\title{Results of pairwise comparisons of simulation scenarios}
\format{
A named list of 12 \code{element}s. See documentation for details.
}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
 \href{https://doi.org/10.1071/PC17002}{DOI:10.1071/PC17002}
}
\description{
Results of pairwise comparisons of simulation scenarios included
in the sensitivity test scenario 'ST_Classic' using \code{pairwise}.
}
\examples{
data("pac.clas.pairw")
head(pac.clas.pairw)
}
