% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.yr}
\alias{pac.yr}
\title{Collated results from Vortex scenarios - Pacioni et al. (2017)}
\format{
A named list of two \code{element}s:
all (8712 obs, 26 var), means (2904 obs, 25 var).
}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017)
Predators and genetic fitness: key threatening factors for the conservation
of bettong species. Pacific Conservation Biology.
\href{https://doi.org/10.1071/PC17002}{DOI:10.1071/PC17002}
}
\description{
Data from Pacioni et al. (2017) used to conduct a sensitivity
analysis on demographic parameters. Vortex outputs, from the project named
'Pacioni_et_al' and (Single-Factor) sensitivity test scenario 'ST_Classic'
(.yr files), were collated with \code{collate_yr}.
}
\examples{
data("pac.yr")
head(pac.yr)
}
