% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pext.R
\name{pval}
\alias{pval}
\title{Calculates p-values from z-values}
\usage{
pval(x)
}
\arguments{
\item{x}{z-values}
}
\value{
A numeric vector of length equal to length(x)
}
\description{
\code{pval} calculates one-tailed p values from a vector that contains
  z-values and it is geenrally used internally.
}
\examples{
 z <- c(1.645, 1.96, 3.09)
 pval(z)
}

