% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vortexr.R
\name{se2sd}
\alias{se2sd}
\title{Standard Error from a vector}
\usage{
se2sd(se, no)
}
\arguments{
\item{se}{A standard error of a set of values}

\item{no}{The number of values}
}
\value{
The standard deviation of the values
}
\description{
Standard Error from a vector
}

