% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexr.R
\name{sta.evy5}
\alias{sta.evy5}
\title{Collated results from Vortex scenarios - Campbell et al (in press)}
\format{a \code{data.frame} with 1020 observations of 47 variables.}
\source{
Campbell et al. In press. Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management
}
\usage{
data(sta.evy5)
}
\description{
A dataset with the results from Vortex scenarios used in
 Campbell et al (in press) to simulate major application of control measures
 in every 5 year cycle. Vortex outputs, from the project named
 "Starlingv3PopBased" and the sensitivity test scenario "MReductEvy5"
 (.stdat files), were collated with \code{collate_dat}.
}

