% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{sta.main}
\alias{sta.main}
\title{Collated results from Vortex scenarios - Campbell et al (2016)}
\format{a \code{data.frame} with 1632 observations of 44 variables.}
\source{
Campbell et al. (2016). Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management, 23, 81-99
}
\usage{
data(sta.main)
}
\description{
A dataset with the results from the main Vortex scenarios used
 in Campbell et al (2016). Vortex outputs, from the project named
 "Starlingv3PopBased" (.dat files), were collated with \code{collate_dat}.
}

