% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vortexr.R
\name{campbell}
\alias{campbell}
\title{Raw Vortex outputs from Campbell et al (in press)}
\format{Several .dat and .stdat files.}
\source{
Campbell et al. In press. Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management
}
\usage{
# To retrieve the path to the files use:
 system.file("extdata", "campbell", package="vortexR")
}
\description{
A folder with Vortex outputs from Campbell et al (in press).
}

