% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.clas.lookup}
\alias{pac.clas.lookup}
\title{Look-up table}
\format{A \code{data.frame} with 24 observations of 8 variables.}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas.lookup)
}
\description{
Data from Pacioni et al. (in prep) - sensitivity test scenario
"ST_Classic" - were used to generate a look-up table
sizes using \code{lookup_table}.
}

