% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexr.R
\name{pac.clas.Ne}
\alias{pac.clas.Ne}
\title{Effective population size}
\format{A \code{data.frame} with 24 observations of 2 variables.}
\source{
Pacioni, C., Spencer, P.B.S., Lacy, R.C., and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas.Ne)
}
\description{
Data from Pacioni et al. (in prep) - sensitivity test scenario
"ST_Classic" - were used to calculate the effective population size
sizes using \code{Ne}.
}

