% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vortexr.R
\name{pac.clas.pairw}
\alias{pac.clas.pairw}
\title{Results of pairwise comparisons of simulation scenarios}
\format{A named list of 12 \code{element}s. See documentation for details.}
\source{
Pacioni, C., Spencer, P.B.S., Lacy, R.C., and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas.pairw)
}
\description{
Results of pairwise comparisons of simulation scenarios included
in the sensitivity test scenario "ST_Classic" using \code{pairwise}.
}

