% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.clas}
\alias{pac.clas}
\title{Collated results from Vortex scenarios - Pacioni et al. (in prep)}
\format{a \code{data.frame} of 2904 observations of 68 variables.}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (in prep)
  Predators and genetic fitness: key threatening factors for the conservation
  of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.clas)
}
\description{
Subset (only 3 runs) of data from Pacioni et al. (in prep) used
  to conduct a sensitivity analysis on demographic parameters. Vortex
  outputs, from the project named "Pacioni_et_al" and (Single-Factor)
  sensitivity test scenario "ST_Classic" (.stdat files), were collated with
  \code{collate_dat}.
}
