% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexR_pkg_data.R
\name{pac.lhs}
\alias{pac.lhs}
\title{Collated results from Vortex scenarios - Pacioni et al. (in prep)}
\format{A \code{data.frame} of 6171 observations of 68 variables.}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.lhs)
}
\description{
Data from Pacioni et al. (in prep) used to conduct a sensitivity
 analysis on demographic parameters. Vortex outputs, from the project named
 "Pacioni_et_al" and (Latin Hypercube Sampling) sensitivity test scenario
 "ST_LHS" (.stdat files), were collated with \code{collate_dat}.
}

