% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexr.R
\name{pac.yr}
\alias{pac.yr}
\title{Collated results from Vortex scenarios - Pacioni et al. (in prep)}
\format{A named list of two \code{element}s:
  all (8712 obs, 26 var), means (2904 obs, 25 var).}
\source{
Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (in prep)
 Predators and genetic fitness: key threatening factors for the conservation
 of bettong species. Pacific Conservation Biology.
}
\usage{
data(pac.yr)
}
\description{
Data from Pacioni et al. (in prep) used to conduct a sensitivity
 analysis on demographic parameters. Vortex outputs, from the project named
 "Pacioni_et_al" and (Single-Factor) sensitivity test scenario "ST_Classic"
 (.yr files), were collated with \code{collate_yr}.
}

