% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{PrefixAndRepeat}
\alias{PrefixAndRepeat}
\title{Return a prefixed and repeated string of character}
\usage{
PrefixAndRepeat(chars, times = 1, prefix = "")
}
\arguments{
\item{chars}{A string of characters (popvalue)}

\item{times}{The number of repetitions (ncolpop), default: 1}

\item{prefix}{A text prefix, default: ""}
}
\description{
Return a prefixed and repeated string of character
}

