% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vortexr.R
\name{sta.main}
\alias{sta.main}
\title{Collated results from Vortex scenarios - Campbell et al (in press)}
\format{a \code{data.frame} with 1632 observations of 44 variables.}
\source{
Campbell et al. In press. Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management
}
\usage{
data(sta.main)
}
\description{
A dataset with the results from the main Vortex scenarios used
 in Campbell et al (in press). Vortex outputs, from the project named
 "Starlingv3PopBased" (.dat files), were collated with \code{collate_dat}.
}

